refactor: extract validation helpers from logfwd-config lib.rs
Move validation functions (validate_host_port, validate_bind_addr,
validate_log_level, validate_endpoint_url, is_loopback_url,
validate_output_endpoint_url) and their unit tests into a dedicated
validation.rs module to keep lib.rs under the 3900-line CI guardrail
when merged with main.
